requirements.txt
mtcnn==0.1.0
Pillow==8.1.0
redisai==1.0.1
matplotlib==2.2.3
albumentations==0.5.2
h5py==2.10.0
numpy==1.18.5
pandas==1.0.1
joblib==0.14.1
ml2rt==0.2.0
flask==1.1.1
flask_restful==0.3.8
mlflow==1.14.1
boto3==1.17.46
psycopg2-binary==2.8.6
python-dotenv==0.17.0
scikit-learn==0.24.1
tensorflow-gpu==1.15.5
prospector==1.3.1
pep257==0.7.0
pep8==1.7.1
pyflakes==2.3.1
pylint==2.8.2
frosted==1.4.1
mccabe==0.6.1
pyroma==3.1
mypy==0.812
